+++ HomeAssistant/Utilities/DeviceUID.h
