+++ Shared/API/Models/Components/Automation.swift
