+++ Shared/API/Models/Components/BinarySensor.swift
@@ -64,6 +64,6 @@ class BinarySensor: Entity {
     }
 
     override var EntityColor: UIColor {
-        return self.State == "on" ? colorWithHexString("#DCC91F", alpha: 1) : self.DefaultEntityUIColor
+        return self.State == "on" ? UIColor.onColor : self.DefaultEntityUIColor
     }
 }
