+++ Shared/API/Models/Components/Climate.swift
