+++ Shared/API/Models/Components/DeviceTracker.swift
