+++ Shared/API/Models/Components/Fan.swift
