+++ Shared/API/Models/Components/GarageDoor.swift
