+++ Shared/API/Models/Components/InputBoolean.swift
@@ -23,6 +23,6 @@ class InputBoolean: Entity {
     }
 
     override var EntityColor: UIColor {
-        return self.State == "on" ? colorWithHexString("#DCC91F", alpha: 1) : self.DefaultEntityUIColor
+        return self.State == "on" ? UIColor.onColor : self.DefaultEntityUIColor
     }
 }
