+++ Shared/API/Models/Components/MediaPlayer.swift
@@ -121,7 +121,7 @@ class MediaPlayer: Entity {
     }
 
     override var EntityColor: UIColor {
-        return self.State == "on" ? colorWithHexString("#DCC91F", alpha: 1) : self.DefaultEntityUIColor
+        return self.State == "on" ? UIColor.onColor : self.DefaultEntityUIColor
     }
 }
