+++ Shared/API/Models/Components/Script.swift
@@ -26,6 +26,6 @@ class Script: Entity {
     }
 
     override var EntityColor: UIColor {
-        return self.State == "on" ? colorWithHexString("#DCC91F", alpha: 1) : self.DefaultEntityUIColor
+        return self.State == "on" ? UIColor.onColor : self.DefaultEntityUIColor
     }
 }
