+++ Shared/API/Models/Components/Sensor.swift
