+++ Shared/API/Models/Components/Thermostat.swift
