+++ Shared/API/Models/LocationHistory.swift
