+++ Shared/Common/Extensions/UIColor+HA.swift
@@ -1,0 +1,15 @@
+//
+//  UIColor+HA.swift
+//  Shared
+//
+//  Created by Stephan Vanterpool on 9/15/18.
+//  Copyright © 2018 Robbie Trencheny. All rights reserved.
+//
+
+import UIKit
+
+extension UIColor {
+    public static let onColor = UIColor(hue: 0.15, saturation: 0.75, brightness: 0.49, alpha: 1.0)
+    public static let defaultEntityColor = UIColor(hue: 0.58, saturation: 0.4, brightness: 0.44, alpha: 1.0)
+    
+}
