+++ Shared/Settings/SettingsStore.swift
@@ -92,6 +92,15 @@ public class SettingsStore {
         }
     }
 
+    public var timezone: String? {
+        get {
+            return prefs.string(forKey: "time_zone")
+        }
+        set {
+            prefs.setValue(newValue, forKey: "time_zone")
+        }
+    }
+
     // MARK: - Private helpers
 
     private var hasMigratedConnection: Bool {
