+++ Shared/Shared-Bridging-Header.h
@@ -1,0 +1,13 @@
+//
+//  Shared-Bridging-Header.h
+//  HomeAssistant
+//
+//  Created by Stephan Vanterpool on 9/15/18.
+//  Copyright © 2018 Robbie Trencheny. All rights reserved.
+//
+
+#ifndef Shared_Bridging_Header_h
+#define Shared_Bridging_Header_h
+
+
+#endif /* Shared_Bridging_Header_h */
